+++ pages/myinfo/bookmark/index.page.tsx
@@ -2,23 +2,11 @@ import React, { useEffect } from 'react';
 
 import { useQueryClient } from '@tanstack/react-query';
 
-import { useDropdownStore } from '@stores/dropdownStore';
-
-import DynamicTechBlogComponent from '@components/features/main/dynamicTechBlogComponent';
-
-import { TechInfiniteDataType } from '@/types/infiniteQueryType';
-
 import MyInfo from '../index.page';
-import { useInfiniteMyInfoBookmark } from './api/useInfiniteMyInfoBookmark';
-import { MyinfoBookmarkDropdownProps } from './bookmarkType';
+import BookmarkSection from './BookmarkSection';
 
 export default function BookMark() {
   const queryClient = useQueryClient();
-  const { sortOption } = useDropdownStore();
-
-  const myInfoBookmarkData = useInfiniteMyInfoBookmark(
-    sortOption as MyinfoBookmarkDropdownProps,
-  ) as TechInfiniteDataType;
 
   useEffect(() => {
     queryClient.invalidateQueries({ queryKey: ['techBlogBookmark'] });
@@ -26,12 +14,7 @@ export default function BookMark() {
 
   return (
     <MyInfo>
-      <DynamicTechBlogComponent
-        data={myInfoBookmarkData}
-        skeletonCnt={10}
-        isScroll={true}
-        type='myinfo'
-      />
+      <BookmarkSection />
     </MyInfo>
   );
 }
